rename calibration_data_parser module to calibration_parser
