+++ tutorials/03_nurbs_surface_reconstruction_tutorial.py
@@ -11,7 +11,7 @@ from artist.core.heliostat_ray_tracer import HeliostatRayTracer
 from artist.core.loss_functions import KLDivergenceLoss
 from artist.core.surface_reconstructor import SurfaceReconstructor
 from artist.io import paint_scenario_parser
-from artist.io.calibration_data_parser import CalibrationDataParser
+from artist.io.calibration_parser import CalibrationDataParser
 from artist.io.paint_calibration_parser import PaintCalibrationDataParser
 from artist.scenario.scenario import Scenario
 from artist.util import config_dictionary, index_mapping, set_logger_config, utils
